rename: BookDto 패키지 변경 application 하위 → domain 하위 (#8)
