+++ bookiary-domain/src/main/java/oz/bookiarybacked/domain/book/dto/BookDto.java
@@ -1,4 +1,4 @@
-package oz.bookiarybacked.application.book.dto;
+package oz.bookiarybacked.domain.book.dto;
 
 import lombok.Builder;
